+++ packages/adapter-node/rollup.config.js
@@ -2,9 +2,31 @@ import { nodeResolve } from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import json from '@rollup/plugin-json';
 import { builtinModules } from 'node:module';
+import { rmSync } from 'node:fs';
 
+/**
+ * @param {string} filepath
+ * @returns {import('rollup').Plugin}
+ */
+function clearOutput(filepath) {
+	return {
+		name: 'clear-output',
+		buildStart: {
+			order: 'pre',
+			sequential: true,
+			handler() {
+				rmSync(filepath, { recursive: true, force: true });
+			}
+		}
+	};
+}
+
+/**
+ * @returns {import('rollup').Plugin}
+ */
 function prefixBuiltinModules() {
 	return {
+		name: 'prefix-built-in-modules',
 		resolveId(source) {
 			if (builtinModules.includes(source)) {
 				return { id: 'node:' + source, external: true };
@@ -20,7 +42,13 @@ export default [
 			file: 'files/index.js',
 			format: 'esm'
 		},
-		plugins: [nodeResolve({ preferBuiltins: true }), commonjs(), json(), prefixBuiltinModules()],
+		plugins: [
+			clearOutput('files/index.js'),
+			nodeResolve({ preferBuiltins: true }),
+			commonjs(),
+			json(),
+			prefixBuiltinModules()
+		],
 		external: ['ENV', 'HANDLER']
 	},
 	{
@@ -29,7 +57,13 @@ export default [
 			file: 'files/env.js',
 			format: 'esm'
 		},
-		plugins: [nodeResolve(), commonjs(), json(), prefixBuiltinModules()],
+		plugins: [
+			clearOutput('files/env.js'),
+			nodeResolve(),
+			commonjs(),
+			json(),
+			prefixBuiltinModules()
+		],
 		external: ['HANDLER']
 	},
 	{
@@ -39,7 +73,13 @@ export default [
 			format: 'esm',
 			inlineDynamicImports: true
 		},
-		plugins: [nodeResolve(), commonjs(), json(), prefixBuiltinModules()],
+		plugins: [
+			clearOutput('files/handler.js'),
+			nodeResolve(),
+			commonjs(),
+			json(),
+			prefixBuiltinModules()
+		],
 		external: ['ENV', 'MANIFEST', 'SERVER', 'SHIMS']
 	},
 	{
@@ -48,6 +88,6 @@ export default [
 			file: 'files/shims.js',
 			format: 'esm'
 		},
-		plugins: [nodeResolve(), commonjs(), prefixBuiltinModules()]
+		plugins: [clearOutput('files/shims.js'), nodeResolve(), commonjs(), prefixBuiltinModules()]
 	}
 ];