+++ packages/kit/src/core/sync/write_types/index.js
@@ -344,9 +344,11 @@ function update_types(config, routes, route, to_delete = new Set()) {
 		}
 
 		if (route.leaf.server) {
-			exports.push('export type PageProps = { data: PageData; form: ActionData }');
+			exports.push(
+				'export type PageProps = { params: RouteParams; data: PageData; form: ActionData }'
+			);
 		} else {
-			exports.push('export type PageProps = { data: PageData }');
+			exports.push('export type PageProps = { params: RouteParams; data: PageData }');
 		}
 	}
 
@@ -413,7 +415,7 @@ function update_types(config, routes, route, to_delete = new Set()) {
 		if (proxies.universal?.modified) to_delete.delete(proxies.universal.file_name);
 
 		exports.push(
-			'export type LayoutProps = { data: LayoutData; children: import("svelte").Snippet }'
+			'export type LayoutProps = { params: LayoutParams; data: LayoutData; children: import("svelte").Snippet }'
 		);
 	}